Move 'feature' tests into their own directory
